backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
python-dotenv>=0.21.0
groq>=0.4.1
pydantic>=1.10.0
python-multipart>=0.0.6
pymavlink>=2.4.37
numpy>=1.21.0
pandas>=1.5.0
requests>=2.31.0 